We were creating authService in both loginModule and homeModule.
We removed providedIn: 'root' from the service and we are declaring it in app.module's providers array so the 2 modules get it from the root Injector.
Also we removed methods calls in auth-service.service's constructor.
We are refreshing the token on header.component's NgOnInit method now.
Cleaned up some debuggers too.